// src/app/models/character/main-statistics.ts
export interface MainStatistics {
    /**
     * Stärke.
     */
    strength: number;

    /**
     * Beweglichkeit.
     */
    agility: number;

    /**
     * Stamina.
     */
    stamina: number;

    /**
     * Magie.
     */
    magic: number;

    /**
     * Geist.
     */
    spirit: number;

    /**
     * Intelligenz.
     */
    intelligence: number;
}